requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
Faker==12.0.1
httpie==3.2.2
idna==3.6
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mixer==7.2.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
